+++ android/app/release/output-metadata.json
@@ -12,7 +12,7 @@
       "filters": [],
       "attributes": [],
       "versionCode": 2,
-      "versionName": "115.1.2",
+      "versionName": "115.1.3",
       "outputFile": "app-release.apk"
     }
   ],
