+++ src-tauri/tauri.conf.json
@@ -8,7 +8,7 @@
   },
   "package": {
     "productName": "RisuAI",
-    "version": "115.1.2"
+    "version": "115.1.3"
   },
   "tauri": {
     "allowlist": {
