+++ src/ts/process/modules.ts
@@ -3,7 +3,8 @@ import { alertConfirm, alertError, alertModuleSelect, alertNormal } from "../ale
 import { DataBase, setDatabase, type customscript, type loreBook, type triggerscript } from "../storage/database"
 import { downloadFile } from "../storage/globalApi"
 import { get } from "svelte/store"
-import { CurrentCharacter, CurrentChat } from "../stores"
+import { CurrentChat } from "../storage/database"
+import { CurrentCharacter } from "../storage/database"
 import { selectSingleFile } from "../util"
 import { v4 } from "uuid"
 import { convertExternalLorebook } from "./lorebook"
