+++ src/ts/process/templates/chatTemplate.ts
@@ -2,7 +2,7 @@ import { Template } from '@huggingface/jinja';
 import type { OpenAIChat } from '..';
 import { get } from 'svelte/store';
 import { DataBase } from 'src/ts/storage/database';
-import { CurrentCharacter } from 'src/ts/stores';
+import { CurrentCharacter } from "src/ts/storage/database";
 
 export const chatTemplates = {
     'llama3': "{% set bos_token = '<|begin_of_text|>' %}{% set loop_messages = messages %}{% for message in loop_messages %}{% set content = '<|start_header_id|>' + message['role'] + '<|end_header_id|>\n\n'+ message['content'] | trim + '<|eot_id|>' %}{% if loop.index0 == 0 %}{% set content = bos_token + content %}{% endif %}{{ content }}{% endfor %}{{ '<|start_header_id|>assistant<|end_header_id|>\n\n' }}",
