+++ src/ts/process/triggers.ts
@@ -3,7 +3,9 @@ import { DataBase, type Chat, type character } from "../storage/database";
 import { tokenize } from "../tokenizer";
 import { getModuleTriggers } from "./modules";
 import { get } from "svelte/store";
-import { CurrentCharacter, CurrentChat, selectedCharID } from "../stores";
+import { selectedCharID } from "../stores";
+import { CurrentChat } from "../storage/database";
+import { CurrentCharacter } from "../storage/database";
 import { processMultiCommand } from "./command";
 import { parseKeyValue } from "../util";
 import { alertError, alertInput, alertNormal, alertSelect } from "../alert";
