+++ version.json
@@ -1,1 +1,1 @@
-{"version":"115.1.2"}
+{"version":"115.1.3"}
